requirements.txt
openai==0.19.0
python-dotenv==0.19.2